// src/app/shared/types/api-types/bsde-project-dto.interface.ts
export interface BsdeProjectDTO {
  projectId?: string;
  projectNumber?: number;
  projectPn?: string;
  projectSn?: string;
  projectFleetId?: string;
  projectName?: string;
  projectDescription?: string;
  projectStartDate?: Date;
  projectEndDate?: Date;
  fromSap?: boolean;
  sapEquipmentId?: string;
  sapProjectDefinition?: string;
  sapProjectProfile?: string;
  sapProjectResponsible?: string;
  sapProjectPlant?: string;
  priorityLevel?: string;
  stdExchangeParam?: string;
  repairCenterCode?: string;
  lockUserLogbook?: string;
  lockSessionLogbook?: string;
  lockUserCm?: string;
  lockSessionCm?: string;
  customerCode?: string;
  statusStateText?: string;
  projectFleetCode?: string;
  projectFleetName?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  taskTypeCode?: string;
  whId?: number;
  cmId?: number;
  cmName?: string;
}